Move fetch and request to peerDependencies
