+++ webpack.config.js
@@ -1,6 +1,6 @@
 module.exports = {
     entry: {
-        restful: './src/restful.js',
+        restful: './src/index.js',
     },
     module: {
         loaders: [{
@@ -9,6 +9,11 @@ module.exports = {
             loader: 'babel-loader',
         }],
     },
+    plugins: [
+        new webpack.DefinePlugin({
+            'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
+        }),
+    ],
     output: {
         path: './dist',
         filename: '[name].js',
